requirements.txt
facebook-sdk
ollama
python-dotenv
requests
python-telegram-bot
apscheduler
pytz
metapub
edge-tts
moviepy>=2.1.2
python-amazon-paapi>=6.2.0
feedparser
Pillow
chromadb
langchain
langchain-community
langchain-huggingface
pypdf
faiss-cpu
sentence-transformers
